Move FeatureSwitches into its own component for ease of testing
