+++ apps/rule-manager/client/src/ts/components/layout/ProfileMenu.tsx
@@ -1,52 +1,12 @@
-import {
-  EuiBasicTable,
-  EuiBasicTableColumn,
-  EuiContextMenu,
-  EuiSwitch,
-} from "@elastic/eui";
-import React, { useContext } from "react";
-import {
-  FeatureSwitch,
-  FeatureSwitchesContext,
-} from "../context/featureSwitches";
-import { PageContext } from "../../utils/window";
+import { EuiContextMenu } from "@elastic/eui";
+import React from "react";
+import { FeatureSwitchTable } from "../FeatureSwitches";
 
 export const ProfileMenu = () => {
-  const { featureSwitches, toggleFeatureSwitch, getFeatureSwitchValue } =
-    useContext(FeatureSwitchesContext);
-
-  const featureColumns: Array<EuiBasicTableColumn<FeatureSwitch>> = [
-    {
-      field: "name",
-      name: "Feature",
-    },
-    {
-      field: "active",
-      name: "Active",
-      render: (_, featureSwitch) =>
-        getFeatureSwitchValue(featureSwitch.id).toString(),
-    },
-    {
-      field: "override",
-      name: "Override",
-      render: (_, featureSwitch) => (
-        <EuiSwitch
-          label=""
-          checked={featureSwitch.overridden}
-          onChange={() => toggleFeatureSwitch(featureSwitch.id)}
-        />
-      ),
-    },
-  ];
   const panels = [
     {
       id: 0,
       items: [
-        {
-          name: "Logout",
-          icon: "exit",
-          onClick: () => (window.location.href = "/logout"),
-        },
         {
           name: "Feature switches",
           icon: "starEmptySpace",
@@ -58,9 +18,7 @@ export const ProfileMenu = () => {
       id: 1,
       title: "Feature switches",
       width: 400,
-      content: (
-        <EuiBasicTable columns={featureColumns} items={featureSwitches} tableLayout="auto" />
-      ),
+      content: FeatureSwitchTable(),
     },
   ];
